feat: cardDetailPage State custom hook 분리
