+++ docs/REQUIREMENTS.md
@@ -17,7 +17,7 @@
 
 - [x] 카드 번호 입력하기
   - [x] 숫자 4자리마다 나눠진다.
-  - [] 뒷 8자리는 \*로 표시된다.
+  - [x] 뒷 8자리는 \*로 표시된다.
   - [x] 총 입력된 숫자가 16자리이다.
   - [] 입력된 숫자가 16자리 미만일 경우
     - [] 안내문구(입력 중간에 다른 입력창으로 넘어갔을 경우에만 확인 가능)
@@ -35,10 +35,10 @@
 
 ## step1 리펙터링 체크리스트
 
-- [] 커스텀 훅 분리
-- [] styled-component 이름 변경
+- [x] 커스텀 훅 분리
+- [x] styled-component 이름 변경
 - [] 컴포넌트 추상화 분리
-  - [] input 추상화
+  - [x] input 추상화
 - [] 입력 값 유효 검사
   - [] 만료기간의 월(1월~12월) 안내
   - [] 카드 번호 16자리 미만 입력 시 안내
