+++ src/component/CardDetailPage/CardDetailForm/CardOwnerNameInput/CardOwnerNameInput.tsx
@@ -14,7 +14,7 @@ function CardOwnerNameInput({
 }: CardOwnerNameInputProps) {
   const props = {
     type: "text",
-    minLength: 30,
+    maxLength: 30,
     isRequired: false,
     placeholder: "카드에 표시된 이름과 동일하게 입력하세요.",
     onInput: changeCardOwnerName,
