+++ src/hooks/useCardOwnerName.tsx
@@ -1,0 +1,14 @@
+import React, { useState } from "react";
+
+function useCardOwnerName() {
+  const [cardOwnerName, setCardOwnerName] = useState("");
+
+  const changeCardOwnerName = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const name = e.target.value;
+    setCardOwnerName(name);
+  };
+
+  return { cardOwnerName, changeCardOwnerName };
+}
+
+export default useCardOwnerName;
